requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
asyncpg>=0.30.0
sqlalchemy>=2.0.0
greenlet>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
aiosqlite>=0.20.0
web3>=7.0.0
